pyproject.toml
[tool.poetry]
name = "mf-bot"
version = "0.1.0"
description = ""
authors = ["Boris Potapov"]
readme = "README.md"
packages = [{include = "mf_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
